+++ InBrainSdk/Assets/Plugins/iOS/InBrainSurveys/Source/InBrainProxyViewController.h
@@ -12,7 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface InBrainProxyViewController : UIViewController<InBrainDelegate>
 
 @property(nonatomic, copy, nullable) void (^onRewardsReceived)(NSString* rewards);
-@property(nonatomic, copy, nullable) void (^onRewardsViewDismissed)();
+@property(nonatomic, copy, nullable) void (^onRewardsViewDismissed)(NSString* dismissedResult);
 
 @property(nonatomic, copy) NSString* surveyId;
 @property(nonatomic, copy) NSString* searchId;